requirements.txt
annotated-types==0.7.0
anyio==4.4.0
dependency-injector==4.41.0
diskcache==5.6.3
fastapi-slim==0.111.1
idna==3.7
Jinja2==3.1.4
llama_cpp_python==0.2.82
MarkupSafe==2.1.5
numpy==2.0.0
pydantic==2.8.2
pydantic_core==2.20.1
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.12.2
